add(fetchEvolutionChain) implemented the function that fetches the evolution chain of the pokemon.
This commit implements the functiont that fetches the evolution chain of a pokemon
Cretaed a type that represents the evolution chaine. THis type is self-referential type
Created and implemented a utility function that creates the photo URL of pokemons based on their id.
ID digit less than 3 are padded with 0s on the start