feat(pokemon card) created the pokemon card
This commit initially adds the pokemon card component that is utilized in the homepage. It displayes pokemon's name, id, photo and types.
Change the component pokemon-type-icon to badge
Edited some color gradient in pokemon card for better ui.
made the pokemon type in lowercase and changed type to types as one of the field of the Pokemon interface