fat(added accent to profile page) accent colors and graphs
This commit adds an accent colors in the profiles page. This depends on the colors of the pokemon
Adds the graph that shows the statistics of the pokemon.

Fixed errors specifically in the states of the component
Added radial component for stats and evolution component for evolution tree

Todo: Create island settins in the profile page.
Todo: Implement the next and previous buttons
Todo: Implement the going home buttons